back/requirements.txt
asgiref==3.2.7
astroid==2.3.3
attrs==19.3.0
autopep8==1.5.1
backcall==0.1.0
bleach==3.1.3
branca==0.3.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.5
django-extensions==2.2.9
djangorestframework==3.11.0
dnspython==1.16.0
drf-yasg==1.17.1
entrypoints==0.3
folium==0.10.1
idna==2.9
importlib-metadata==1.5.2
inflection==0.3.1
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
itypes==1.1.0
jedi==0.16.0
Jinja2==2.11.1
jsonschema==3.2.0
jupyter-client==6.1.1
jupyter-core==4.6.3
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.1.2
mccabe==0.6.1
mistune==0.8.4
mysql-connector-python==8.0.19
mysqlclient==1.4.6
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.17.3
packaging==20.3
pandas==0.25.3
pandocfilters==1.4.2
parso==0.6.2
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==3.0.4
protobuf==3.6.1
pycodestyle==2.5.0
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.4.4
pyparsing==2.4.6
pyrsistent==0.16.0
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
pywin32==227
pywinpty==0.5.7
pyzmq==19.0.0
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scipy==1.3.1
seaborn==0.9.0
Send2Trash==1.5.0
six==1.14.0
sqlparse==0.3.1
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
wcwidth==0.1.9
webencodings==0.5.1
wrapt==1.11.2
zipp==3.1.0
